Float ->Float64

+++ src/atom.jl
@@ -178,7 +178,7 @@ end
 # ======================== Grid(N, h, r0, r, r′ , k, am) ===============
 
 @doc raw"""
-    Grid(N::Int, h::Float64, r0::Float64, r::Vector{Float64}, r′::Vector{Float64}, k::Int, am::Vector{Float})
+    Grid(N::Int, h::Float64, r0::Float64, r::Vector{Float64}, r′::Vector{Float64}, k::Int, am::Vector{Float64})
 
 Type to specify the `Grid` on which the atomic wavefunction is defined, with fields `N::Int` (number of grid points),
 `h::Float64` (step size on uniform grid), `r0::Float64` (scale factor for physical grid in a.u.),
